IA/requirement.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
timm>=0.9.0